loading-sessions-automate
Add checkTmuxSessionExists() method to BaseSession
Add helper functions for bash commands so I can await them to avoid race
conditions.
LoadSessions will now create a temp session so we don't have to be in
tmux when we run the script, session will be deleted at the end and we
will attach to the first session in the list.