gitRepoCloning
Add gitRepoLink to Window and Pane
Will now check each folder individually and if a folder is missing it
will check for gitRepoLink on the Pane, if there is a repo link then it
will clone it and navigate to the folder again.
Fixed printing sessions with 'ls', was not setting current sessions.
Some refactoring