sessionNamesAndUserPrompt
Add ability to name sessions with a prompt and a default name of the
name of the branch currently in, if no branch then user will be asked to
pick a name.

Add a select menu when loading also user input will search through the
list and shorten it based on input.