conflicts
Add conflict resolution automation

When user types 'kra git conflict-handle' it will prompt with a list
with conflicting files
Selecting a file will open it in neovim and split the screen
Once handled and file is closed file will be grepped for conflict
markers, if none found it will remove file from list and prompt again,
if markers are found it will prompt again with the file still in the
list.

Once all conflicts are handled it will print a table with all the files
resolved.